finish poc

+++ invman_server/lib/src/stock/data/stock_client.dart
@@ -3,5 +3,5 @@ import 'package:invman_server/src/generated/protocol.dart';
 abstract interface class StockClient {
   Future<List<Stock>> search({required String query, int limit = 10});
   Future<Stock> get({required String symbol});
-  Future<int> currencyChange({required String from, required String to});
+  Future<double> currencyChange({required String from, required String to});
 }
